allow tests to run in isolated main processes
spawn rpc processes per environment (node/browser)

run node and main process tests in isolated instances

refactor renderer process to extract common methods

update documentation
