+++ package.json
@@ -42,7 +42,7 @@
     "lint": "eslint .",
     "test": "jest -w 3",
     "test-all": "yarn flow && yarn test && yarn lint",
-    "gen-rpc": "node packages/rpc/bin/cli.js packages/*/src/rpc/{JestWorkerRPC,NuclideE2ERPC}.js"
+    "gen-rpc": "node packages/rpc/bin/cli.js packages/*/src/rpc/{JestWorkerRPC,JestWorkerRPCMain,NuclideE2ERPC}.js"
   },
   "dependencies": {}
 }
