+++ packages/electron/src/rpc/JestWorkerRPCMain.js
@@ -1,0 +1,46 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+import type {TestResult} from '@jest-runner/core/types';
+import type {IPCTestData} from '../../types';
+import runTest from 'jest-runner/build/run_test';
+import docblock from '@jest-runner/core/docblock';
+import fs from 'fs';
+
+import {buildFailureTestResult} from '@jest-runner/core/utils';
+
+import {getResolver} from '../utils/resolver';
+
+const _runTest = (testData: IPCTestData): Promise<TestResult> => {
+  try {
+    return runTest(
+      testData.path,
+      testData.globalConfig,
+      testData.config,
+      getResolver(testData.config, testData.rawModuleMap),
+    );
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.error(error);
+    return buildFailureTestResult(
+      testData.path,
+      error,
+      testData.config,
+      testData.globalConfig,
+    );
+  }
+};
+
+module.exports = {
+  runTest(testData: IPCTestData): Promise<TestResult> {
+    return _runTest(testData);
+  },
+  shutDown(): Promise<any> {
+    return Promise.resolve();
+  },
+};
