+++ packages/electron/src/utils/once.js
@@ -1,0 +1,9 @@
+export const once = fn => {
+  let hasBeenCalled = false;
+  return (...args) => {
+    if (!hasBeenCalled) {
+      hasBeenCalled = true;
+      return fn(...args);
+    }
+  };
+};
